+++ back/dao/src/main/java/genealogy/visualizer/service/impl/LocalityDAOImpl.java
@@ -8,9 +8,11 @@ import genealogy.visualizer.entity.Marriage;
 import genealogy.visualizer.entity.Person;
 import genealogy.visualizer.repository.ChristeningRepository;
 import genealogy.visualizer.repository.DeathRepository;
+import genealogy.visualizer.repository.GodParentRepository;
 import genealogy.visualizer.repository.LocalityRepository;
 import genealogy.visualizer.repository.MarriageRepository;
 import genealogy.visualizer.repository.PersonRepository;
+import genealogy.visualizer.repository.WitnessRepository;
 import genealogy.visualizer.service.LocalityDAO;
 import genealogy.visualizer.service.helper.RepositoryEasyModelHelper;
 import jakarta.persistence.EntityManager;
@@ -33,6 +35,8 @@ public class LocalityDAOImpl implements LocalityDAO {
     private final DeathRepository deathRepository;
     private final PersonRepository personRepository;
     private final MarriageRepository marriageRepository;
+    private final WitnessRepository witnessRepository;
+    private final GodParentRepository godParentRepository;
     private final EntityManager entityManager;
 
     public LocalityDAOImpl(LocalityRepository localityRepository,
@@ -40,12 +44,16 @@ public class LocalityDAOImpl implements LocalityDAO {
                            DeathRepository deathRepository,
                            PersonRepository personRepository,
                            MarriageRepository marriageRepository,
+                           WitnessRepository witnessRepository,
+                           GodParentRepository godParentRepository,
                            EntityManager entityManager) {
         this.localityRepository = localityRepository;
         this.christeningRepository = christeningRepository;
         this.deathRepository = deathRepository;
         this.personRepository = personRepository;
         this.marriageRepository = marriageRepository;
+        this.witnessRepository = witnessRepository;
+        this.godParentRepository = godParentRepository;
         this.entityManager = entityManager;
     }
 
@@ -56,21 +64,23 @@ public class LocalityDAOImpl implements LocalityDAO {
 
     @Override
     @Transactional(isolation = Isolation.SERIALIZABLE)
-    public void delete(Long id) {
+    public void delete(Long id) throws IllegalArgumentException {
+        if (id == null)
+            throw new IllegalArgumentException("Cannot delete locality without id");
         christeningRepository.updateLocalityId(id, null);
-        christeningRepository.updateGodParentLocalityId(id, null);
+        godParentRepository.updateGodParentLocalityId(id, null);
         deathRepository.updateLocalityId(id, null);
         personRepository.updateBirthLocalityId(id, null);
         personRepository.updateDeathLocalityId(id, null);
         marriageRepository.updateWifeLocalityId(id, null);
         marriageRepository.updateHusbandLocalityId(id, null);
-        marriageRepository.updateWitnessLocalityId(id, null);
+        witnessRepository.updateWitnessLocalityId(id, null);
         localityRepository.deleteById(id);
     }
 
     @Override
     @Transactional(isolation = Isolation.SERIALIZABLE)
-    public Locality save(Locality locality) {
+    public Locality save(Locality locality) throws IllegalArgumentException, EmptyResultDataAccessException {
         if (locality.getId() != null)
             throw new IllegalArgumentException("Cannot save locality with id");
         Locality localityForSave = locality.clone();
@@ -89,25 +99,25 @@ public class LocalityDAOImpl implements LocalityDAO {
 
     @Override
     @Transactional(isolation = Isolation.SERIALIZABLE)
-    public Locality update(Locality locality) {
-        if (locality.getId() == null)
+    public Locality update(Locality locality) throws IllegalArgumentException, EmptyResultDataAccessException {
+        Long id = locality.getId();
+        if (id == null)
             throw new IllegalArgumentException("Cannot update locality without id");
-        Locality updatedLocality = localityRepository.update(locality);
-        if (updatedLocality == null)
-            throw new EmptyResultDataAccessException("Updating locality failed", 1);
-        localityRepository.deleteAnotherNamesById(updatedLocality.getId());
+        Locality existInfo = this.findFullInfoById(id);
+        localityRepository.update(locality).orElseThrow(() -> new EmptyResultDataAccessException("Updating locality failed", 1));
+        localityRepository.deleteAnotherNamesById(id);
         if (locality.getAnotherNames() != null) {
-            locality.getAnotherNames().forEach(an -> localityRepository.insertAnotherName(updatedLocality.getId(), an));
+            locality.getAnotherNames().forEach(an -> localityRepository.insertAnotherName(id, an));
         }
-        updateLinks(updatedLocality, locality);
+        updateLinks(existInfo, locality);
         entityManager.flush();
         entityManager.clear();
-        return this.findFullInfoById(updatedLocality.getId());
+        return this.findFullInfoById(id);
     }
 
     @Override
     @Transactional(readOnly = true)
-    public Locality findFullInfoById(Long id) {
+    public Locality findFullInfoById(Long id) throws EmptyResultDataAccessException {
         String errorMes = String.format("Locality not found by id: %d", id);
         localityRepository.findWithDeaths(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
         localityRepository.findWithAnotherNames(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
@@ -115,12 +125,14 @@ public class LocalityDAOImpl implements LocalityDAO {
         localityRepository.findWithPersonsWithBirthLocality(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
         localityRepository.findWithPersonsWithDeathLocality(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
         localityRepository.findWithMarriagesWithWifeLocality(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
+        localityRepository.findWithWitnesses(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
+        localityRepository.findWithGodParents(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
         return localityRepository.findWithMarriagesWithHusbandLocality(id).orElseThrow(() -> new EmptyResultDataAccessException(errorMes, 1));
     }
 
     @Override
     @Transactional(readOnly = true)
-    public List<Locality> filter(LocalityFilterDTO filter) {
+    public List<Locality> filter(LocalityFilterDTO filter) throws EmptyResultDataAccessException {
         CriteriaBuilder cb = entityManager.getCriteriaBuilder();
         CriteriaQuery<Locality> cq = cb.createQuery(Locality.class);
         Root<Locality> root = cq.from(Locality.class);
@@ -135,7 +147,11 @@ public class LocalityDAOImpl implements LocalityDAO {
             predicates.add(cb.equal(cb.lower(root.get("type")), filter.getType().getName().toLowerCase()));
         }
         cq.select(root).where(predicates.toArray(new Predicate[0]));
-        return entityManager.createQuery(cq).getResultList();
+        List<Locality> result = entityManager.createQuery(cq).getResultList();
+        if (result == null || result.isEmpty()) {
+            throw new EmptyResultDataAccessException(String.format("Localities not found filter: %s", filter), 1);
+        }
+        return result;
     }
 
     @Transactional(isolation = Isolation.SERIALIZABLE)
